create ZickZack.cs
